# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DroneModelEnum": {
				"enum": [
					"Lightweight",
					"Middleweight",
					"Cruiserweight",
					"Heavyweight"
				],
				"type": "string"
			},
			"DroneStateEnum": {
				"enum": [
					"IDLE",
					"LOADING",
					"LOADED",
					"DELIVERING",
					"DELIVERED",
					"RETURNING"
				],
				"type": "string"
			},
			"GetDroneDto": {
				"properties": {
					"serialNumber": {
						"type": "string"
					},
					"model": {
						"$ref": "#/components/schemas/DroneModelEnum"
					},
					"weightLimit": {
						"type": "number",
						"format": "double"
					},
					"batteryCapacity": {
						"type": "number",
						"format": "double"
					},
					"state": {
						"$ref": "#/components/schemas/DroneStateEnum"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"batteryCapacity",
					"id",
					"model",
					"serialNumber",
					"state",
					"weightLimit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateDroneDto": {
				"properties": {
					"batteryCapacity": {
						"type": "number",
						"format": "double"
					},
					"model": {
						"$ref": "#/components/schemas/DroneModelEnum"
					},
					"serialNumber": {
						"type": "string"
					},
					"state": {
						"$ref": "#/components/schemas/DroneStateEnum"
					},
					"weightLimit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"batteryCapacity",
					"model",
					"serialNumber",
					"state",
					"weightLimit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Medicament": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "string"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"weight",
					"code",
					"image"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Drone": {
				"properties": {
					"id": {
						"type": "string"
					},
					"serialNumber": {
						"type": "string"
					},
					"model": {
						"$ref": "#/components/schemas/DroneModelEnum"
					},
					"weightLimit": {
						"type": "number",
						"format": "double"
					},
					"batteryCapacity": {
						"type": "number",
						"format": "double"
					},
					"state": {
						"$ref": "#/components/schemas/DroneStateEnum"
					}
				},
				"required": [
					"serialNumber",
					"model",
					"weightLimit",
					"batteryCapacity",
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetBatteryLevelDto": {
				"properties": {
					"serialNumber": {
						"type": "string"
					},
					"model": {
						"$ref": "#/components/schemas/DroneModelEnum"
					},
					"weightLimit": {
						"type": "number",
						"format": "double"
					},
					"batteryCapacity": {
						"type": "number",
						"format": "double"
					},
					"state": {
						"$ref": "#/components/schemas/DroneStateEnum"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"batteryCapacity"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "dispatch_controller",
		"version": "1.0.0",
		"description": "Med Aid Drone Dispatch Controller",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "lazaroofarrill ",
			"email": "lazaroofarrill@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/drones": {
			"post": {
				"operationId": "RegisterDrone",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetDroneDto"
								}
							}
						}
					}
				},
				"description": "Registers a drone in the system",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDroneDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "CheckAvailableDrones",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Drone"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/drones/{droneId}/items/{medicamentId}": {
			"post": {
				"operationId": "LoadItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Loads a drone with an instance of a medicament",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "droneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "medicamentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/drones/{droneId}/items": {
			"get": {
				"operationId": "CheckLoadedItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"quantity": {
												"type": "number",
												"format": "double"
											},
											"medicament": {
												"$ref": "#/components/schemas/Medicament"
											}
										},
										"required": [
											"quantity",
											"medicament"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "droneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/drones/{droneId}/battery": {
			"get": {
				"operationId": "CheckBatteryLevel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetBatteryLevelDto"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "droneId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}